fix(std): close CI compile holes and leftover fail-open paths
Host clang -Werror, missing StdLibTests link deps, and the linux-arm64
assertion failures are fixed; the audit also closed wrap, injection,
and protocol holes that still failed open.

+++ std/src/image/gif/gif.c
@@ -208,7 +208,7 @@ int neverc_gif_encode(const neverc_gif_frame_t *frame,
         (int)frame->transparent_index >= frame->palette_size)
         return -1;
     if ((uint64_t)frame->width * frame->height > GIF_MAX_PIXELS) return -1;
-    if (frame->height > SIZE_MAX / frame->width) return -1;
+    if ((size_t)frame->height > SIZE_MAX / (size_t)frame->width) return -1;
     size_t pixel_count = (size_t)frame->width * frame->height;
     if (pixel_count > SIZE_MAX - 1024) return -1;
     /* Indices outside the palette encode as illegal LZW codes (or decode as
@@ -309,7 +309,7 @@ int neverc_gif_decode(const uint8_t *data, size_t len, neverc_gif_image_t *img) 
     pos++; /* pixel aspect ratio */
     if (width == 0 || height == 0 ||
         (uint64_t)width * height > GIF_MAX_PIXELS ||
-        height > SIZE_MAX / width)
+        (size_t)height > SIZE_MAX / (size_t)width)
         return -1;
 
     int has_gct = (packed >> 7) & 1;
@@ -711,7 +711,7 @@ int neverc_gif_frame_to_rgba(const neverc_gif_frame_t *frame,
         (int)frame->transparent_index >= frame->palette_size)
         return -1;
     if ((uint64_t)frame->width * frame->height > GIF_MAX_PIXELS) return -1;
-    if (frame->height > SIZE_MAX / 4 / frame->width) return -1;
+    if ((size_t)frame->height > SIZE_MAX / 4 / (size_t)frame->width) return -1;
 
     size_t npixels = (size_t)frame->width * frame->height;
     size_t nbytes = npixels * 4u;
@@ -1000,8 +1000,8 @@ int neverc_gif_from_rgba(const uint8_t *rgba, uint32_t width, uint32_t height,
     if (!rgba || !frame || width == 0 || height == 0 ||
         width > UINT16_MAX || height > UINT16_MAX ||
         (uint64_t)width * height > GIF_MAX_PIXELS ||
-        width > SIZE_MAX / 4 ||
-        height > SIZE_MAX / 4 / width)
+        (size_t)width > SIZE_MAX / 4 ||
+        (size_t)height > SIZE_MAX / 4 / (size_t)width)
         return -1;
     memset(frame, 0, sizeof(*frame));
     frame->width = width;
